# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.InvoiceResponse:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      dueDate:
        type: string
      id:
        type: string
      invoiceNumber:
        type: string
      paidAt:
        type: string
      paymentId:
        type: string
      sentAt:
        type: string
      status:
        type: string
      subscriptionId:
        type: string
      tenantId:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.PayInvoiceRequest:
    properties:
      paymentId:
        type: string
      paymentMethod:
        type: string
      transactionId:
        type: string
    type: object
...
  dto.SendInvoiceRequest:
    properties:
      email:
        type: string
    type: object
...
  model.Invoice:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      dueDate:
        type: string
      id:
        type: string
      invoiceNumber:
        type: string
      paidAt:
        type: string
      paymentId:
        type: string
      sentAt:
        type: string
      status:
        $ref: '#/definitions/model.InvoiceStatus'
      subscriptionId:
        type: string
      tenantId:
        type: string
      updatedAt:
        type: string
    type: object
  model.InvoiceStatus:
    enum:
    - draft
    - sent
    - paid
    - overdue
    - canceled
    type: string
    x-enum-varnames:
    - InvoiceStatusDraft
    - InvoiceStatusSent
    - InvoiceStatusPaid
    - InvoiceStatusOverdue
    - InvoiceStatusCanceled
...
host: localhost:8081
info:
  contact: {}
  description: Human Resource Information System API
  title: HRIS API
  version: "1.0"
paths:
...
  /invoices:
    get:
      consumes:
      - application/json
      description: Get all invoices
      parameters:
      - description: Invoice status
        in: query
        name: status
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.InvoiceResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all invoices
      tags:
      - invoices
    post:
      consumes:
      - application/json
      description: Create a new invoice (Admin only)
      parameters:
      - description: Invoice data
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/model.Invoice'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new invoice
      tags:
      - invoices
  /invoices/{id}:
    get:
      consumes:
      - application/json
      description: Get invoice by ID
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get invoice by ID
      tags:
      - invoices
    put:
      consumes:
      - application/json
      description: Update an invoice (Admin only)
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      - description: Invoice data
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/model.Invoice'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an invoice
      tags:
      - invoices
  /invoices/{id}/pay:
    post:
      consumes:
      - application/json
      description: Pay an invoice
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      - description: Payment details
        in: body
        name: pay
        required: true
        schema:
          $ref: '#/definitions/dto.PayInvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Pay an invoice
      tags:
      - invoices
  /invoices/{id}/send:
    post:
      consumes:
      - application/json
      description: Send an invoice (Admin only)
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      - description: Send details
        in: body
        name: send
        required: true
        schema:
          $ref: '#/definitions/dto.SendInvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send an invoice
      tags:
      - invoices
  /invoices/subscription/{subscriptionId}:
    get:
      consumes:
      - application/json
      description: Get invoices by subscription ID
      parameters:
      - description: Subscription ID
        in: path
        name: subscriptionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.InvoiceResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get invoices by subscription ID
      tags:
      - invoices
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"